requirements.txt
# Core dependencies with GPU CUDA 12.1 support
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.0+cu121
torchvision==0.17.0+cu121

# Main AI model and vision processing
transformers==4.49.0
einops>=0.7.0
timm>=0.9.12
safetensors>=0.4.0

# Backend API and operational libraries
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=10.2.0
tqdm>=4.66.0
hf-xet>=1.4.0
